chore: letter 패키지 수정과 충돌 해결 및 커밋
