+++ backend/src/main/java/com/now/naaga/common/exception/ControllerExceptionHandler.java
@@ -9,7 +9,7 @@ import org.springframework.web.HttpMediaTypeNotSupportedException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-`import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
+import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
 @RestControllerAdvice
 public class ControllerExceptionHandler {
@@ -29,8 +29,7 @@ public class ControllerExceptionHandler {
     @ExceptionHandler({
             HttpMessageNotReadableException.class,
             MethodArgumentNotValidException.class,
-            HttpMediaTypeNotSupportedException.class,
-            MethodArgumentTypeMismatchException.class})
+            HttpMediaTypeNotSupportedException.class})
     public ResponseEntity<ExceptionResponse> handleTypeMismatchException(final Exception e) {
         final CommonExceptionType commonExceptionType = CommonExceptionType.INVALID_REQUEST_BODY;
         final ExceptionResponse exceptionResponse = new ExceptionResponse(commonExceptionType.errorCode(), commonExceptionType.errorMessage());
@@ -39,7 +38,17 @@ public class ControllerExceptionHandler {
 
         return ResponseEntity.status(commonExceptionType.httpStatus()).body(exceptionResponse);
     }
-
+    
+    @ExceptionHandler(MethodArgumentTypeMismatchException.class)
+    public ResponseEntity<ExceptionResponse> handleArgumentTypeMismatchException(final Exception e) {
+        final CommonExceptionType commonExceptionType = CommonExceptionType.INVALID_REQUEST_PARAMETERS;
+        final ExceptionResponse exceptionResponse = new ExceptionResponse(commonExceptionType.errorCode(), commonExceptionType.errorMessage());
+        
+        log.info("error = {}", exceptionResponse);
+        
+        return ResponseEntity.status(commonExceptionType.httpStatus()).body(exceptionResponse);
+    }
+    
     @ExceptionHandler(InternalException.class)
     public ResponseEntity<ExceptionResponse> handleInternalException(final InternalException e) {
         final BaseExceptionType internalExceptionType = e.exceptionType();
