+++ backend/src/test/java/com/now/naaga/letter/application/LetterServiceTest.java
@@ -39,6 +39,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.*;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @ActiveProfiles("test")
@@ -63,7 +64,107 @@ class LetterServiceTest {
 
     @Autowired
     private LetterBuilder letterBuilder;
-
+    
+    @Test
+    void 이미_읽은_쪽지의_경우_읽은_쪽지_로그를_기록하지_않는다() {
+        // given
+        final Player player = playerBuilder.init()
+                                           .build();
+        
+        final Place destination = placeBuilder.init()
+                                              .position(잠실_루터회관_정문_좌표)
+                                              .build();
+        
+        final Game game = gameBuilder.init()
+                                     .place(destination)
+                                     .player(player)
+                                     .startPosition(잠실역_교보문고_좌표)
+                                     .build();
+        
+        final Player letterRegister = playerBuilder.init()
+                                                   .build();
+        
+        final Letter letter = letterBuilder.init()
+                                           .registeredPlayer(letterRegister)
+                                           .build();
+        
+        LetterService mockReadLetterService = mock(LetterService.class);
+        
+        // when
+        mockReadLetterService.findLetter(new LetterReadCommand(player.getId(), letter.getId()));
+        mockReadLetterService.findLetter(new LetterReadCommand(player.getId(), letter.getId()));
+        
+        //then
+        verify(mockReadLetterService, atLeast(1)).findLetter(new LetterReadCommand(player.getId(), letter.getId()));
+    }
+    
+    @Test
+    void 읽은_쪽지_로그를_정상적으로_기록한다() {
+        // given
+        final Player player = playerBuilder.init()
+                                           .build();
+        
+        final Place destination = placeBuilder.init()
+                                              .position(잠실_루터회관_정문_좌표)
+                                              .build();
+        
+        final Game game = gameBuilder.init()
+                                     .place(destination)
+                                     .player(player)
+                                     .startPosition(잠실역_교보문고_좌표)
+                                     .build();
+        
+        final Player letterRegister = playerBuilder.init()
+                                                   .build();
+        
+        final Letter letter = letterBuilder.init()
+                                           .registeredPlayer(letterRegister)
+                                           .build();
+        
+        // when
+        letterService.findLetter(new LetterReadCommand(player.getId(), letter.getId()));
+        
+        // then
+        final List<ReadLetterLog> actual = readLetterLogRepository.findAll();
+        final long expected = actual.get(0).getLetter().getId();
+        
+        assertSoftly(softAssertions -> {
+            softAssertions.assertThat(actual).hasSize(1);
+            softAssertions.assertThat(expected).isEqualTo(letter.getId());
+        });
+    }
+    
+    
+    @Test
+    void 읽은쪽지로그에_데이터저장시_진행중인_게임이없으면_예외가_발생한다() {
+        // given && when
+        final Player player = playerBuilder.init()
+                                           .build();
+        
+        final Place destination = placeBuilder.init()
+                                              .position(잠실_루터회관_정문_좌표)
+                                              .build();
+        
+        final Game game = gameBuilder.init()
+                                     .place(destination)
+                                     .player(player)
+                                     .startPosition(잠실역_교보문고_좌표)
+                                     .gameStatus(GameStatus.DONE)
+                                     .build();
+        
+        final Player letterRegister = playerBuilder.init()
+                                                   .build();
+        
+        final Letter letter = letterBuilder.init()
+                                           .registeredPlayer(letterRegister)
+                                           .build();
+        
+        //then
+        final GameException gameException = assertThrows(
+                GameException.class, () -> letterService.findLetter(new LetterReadCommand(player.getId(), letter.getId())));
+        assertThat(gameException.exceptionType()).isEqualTo(NOT_EXIST_IN_PROGRESS);
+    }
+    
     @Transactional
     @Test
     void 쪽지를_정상적으로_생성하고_게임중_등록한_쪽지를_기록으로_남긴다() {
